NGS-general: relocate 'manage_seqs.py' into bcftbx.cli.
