+++ bin/manage_seqs.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python
+#
+#     manage_seqs.py: Utility for handling sets of named sequences
+#     Copyright (C) University of Manchester 2021 Peter Briggs
+#
+from bcftbx.cli.manage_seqs import main
+if __name__ == "__main__":
+     main()
